Made a separate method to create a new bank account, in case the other one wants to be used for something else. CreateAccount() lets the user input name of the account, currency and how much the initial deposit is gonna be.
Added more to the TakeLoan method so the user can input an amount to loan, and limited it to 5 times the users total balance.

The code has yet to be tested. And I haven't done error handling on the user inputs. But I think the logic should be more or less correct.